Move flake scripts out from `overlays`
The flake scripts themselves are fine and do presently
work with the current merge through the use of overlays.

As I've learnt, this can be problematic however.
View the following for reference of what is explained below:
<https://discourse.nixos.org/t/why-does-overlaying-check-with-nodepackages-prettier-cause-a-massive-rebuild/55810>

When overlaying packages, if there is an existing package with the
same name, this should either be avoided altogether or done
with consideration to ensure it is a proper replacement.
`check` for example, is an existing package in `nixpkgs`.
Here a `check` overlay is defined, substituting it
with the `check` script derivation defined in `flake.nix`. With
the current batch of packages in the environment, nothing is affected
fortunately. However, adding `nodePackages.prettier` will cause
a massive rebuild by Nix as there are many packages that are dependent
on the original `check` package. Nix, doing the right thing, works to
build what it needs. Although the final result will be what is expected,
this is more a side effect.

What is the point being argued here if the current implementation is
working?

The big point is that this is a bad practice, since this is
not what overlays were meant for. The same can be achieved by defining
the packages anywhere else, and then referencing them as needed,
rather than through overlays, which can cause unintended side effects,
such as the one described.

Considering this is a repo that is very helpful particularly for
beginners that may new or not as experienced with the Nix ecosystem,
presenting potential beginners to incorrect usage of overlays, can
lead them to doing the same. Hence, the change here would be better
in line with correct implementation. Someone visiting the repo
hoping to implement helper script packages like here, could then
reference this more "correct" implementation. Such as myself,
who ended up following the previous implementation until I encountered
the side effect explained in the post linked above.

The script derivations can really be defined anywhere, such as in a
simple `let` binding, which I've done here. They will have gcroots
linked so they won't be garbage collected, just like in the previous
implementation.

In the end, the resulting behavior remains entirely the same, just
the implementation is corrected to avoid side effects and serve
as an example of better practices.