haxe: Use `haxe_4_0` instead of `haxe` by default
This is to get around `mbedtls_2` being marked as insecure, which is
required for Haxe versions >=4.1. The reality is that this is failing the
PR checks, which is annoying, so this solves the issue for the moment.
There is a comment explaining this so the user can take action as they
see fit until the upstream expression is updated.